feat(storage): seperate intra l0 picker
